Adding Build
Note that the result of the build is different from debug, notably the behavour of the opacity of the buttons and distance of the card to be registred as "swaped".
Probably comes from the fact that the build take scree width/height from pc and debug from mobile. (Game was intended to be mobile)